requirements.txt
# Required dependancies for the project
# To install, run: pip install -r requirements.txt
macaddress
numpy
paramiko
pandas
scipy
scp
plotly
python-dotenv
requests